Extract tool card views
