+++ docs/CODE_QUALITY_AUDIT.md
@@ -43,7 +43,7 @@ The architecture is moving in the right direction: core state is value typed, pe
 
 1. Extract workspace command execution from `WorkspaceModel`.
 2. Extract automation runners from `WorkspaceModel`.
-3. Continue splitting `WorkspaceSwiftUIView` into pane/control files matching the surface structs. The composer, model picker, top bar, sidebar, review pane, design primitives, and transcript message bubbles are now extracted; the next target is the tool-card/artifact-preview family.
+3. Continue splitting `WorkspaceSwiftUIView` into pane/control files matching the surface structs. The composer, model picker, top bar, sidebar, review pane, design primitives, transcript message bubbles, and tool-card/artifact-preview family are now extracted; the next targets are settings, runtime issue panels, terminal/browser panes, and workspace command execution.
 4. Move desktop menu-bar/OAuth orchestration out of `Sources/quill-code-desktop/main.swift`.
 5. Keep the parity matrix updated whenever a feature moves from planned to implemented.
 
@@ -138,3 +138,15 @@ Interface polish changes:
 - Message action controls keep the shared 40 pt minimum hit target and `0.96` press feedback in one focused file.
 - The transcript copy button is now shared from the transcript-message component file, so message bubbles and tool cards do not need separate copy affordance implementations.
 - The workspace shell shrank by another focused chunk, reducing the risk that future transcript edits accidentally touch terminal, settings, or browser panes.
+
+## 2026-06-22 Tool Card Refactor Pass
+
+Overall grade after this slice: **A- foundation, B+ product surface maturity**.
+
+Tool cards and artifact previews moved out of `WorkspaceSwiftUIView.swift` into `QuillCodeToolCardView.swift`. Tool status badges, execution-context chips and rails, artifact chips, document/image/text previews, and raw JSON detail blocks now live beside the tool-card renderer. The workspace shell places transcript timeline items and wires copy actions, but it no longer owns the tool-card rendering family.
+
+Interface polish changes:
+
+- Tool-card header density, status rails, and bounded raw details preserve the existing rhythm while making future polish safer to localize.
+- Artifact chips and previews preserve 40 pt minimum hit areas, pure-white image outline behavior through the design system, and bounded raw JSON/details.
+- The shared transcript copy button is reused from transcript message controls so message and tool-card copy affordances stay consistent.
